requirements.txt
APScheduler==3.9.1
asgiref==3.4.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
blinker==1.4
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==6.7
cloud-init==23.1.2
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Django==3.2.20
django-apscheduler==0.6.2
django-extensions==3.2.3
djangorestframework==3.12.4
google-auth==2.13.0
httplib2==0.9.2
hyperlink==17.3.1
idna==3.4
importlib-resources==5.4.0
incremental==16.10.1
iperf3==0.1.11
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kubernetes==21.7.0
language-selector==0.1
MarkupSafe==1.0
netifaces==0.10.4
oauthlib==3.2.2
PAM==0.4.2
pexpect==4.2.1
pika==1.3.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
PyNaCl==1.4.0
pyOpenSSL==17.5.0
pyserial==3.4
python-apt
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.25
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
rsa==4.9
SecretStorage==2.3.1
service-identity==16.0.0
six==1.16.0
sos
sqlparse==0.4.2
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
typing-extensions==4.1.1
tzdata==2022.5
tzlocal==4.2
ubuntu-advantage-tools
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
virtualenv==15.1.0
websocket-client==1.3.1
zipp==3.6.0
zope.interface==4.3.2
